+++ src/codegen/expression/arithmetic_translator.cpp
@@ -10,7 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "codegen/arithmetic_translator.h"
+#include "codegen/expression/arithmetic_translator.h"
 
 #include "expression/operator_expression.h"
