+++ src/codegen/expression/case_translator.cpp
@@ -10,7 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "codegen/case_translator.h"
+#include "codegen/expression/case_translator.h"
 
 #include "codegen/compilation_context.h"
 #include "codegen/lang/if.h"
