+++ src/codegen/expression/comparison_translator.cpp
@@ -10,7 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "codegen/comparison_translator.h"
+#include "codegen/expression/comparison_translator.h"
 
 #include "expression/comparison_expression.h"
