+++ src/codegen/expression/constant_translator.cpp
@@ -10,7 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "codegen/constant_translator.h"
+#include "codegen/expression/constant_translator.h"
 
 #include "codegen/type/sql_type.h"
 #include "expression/constant_value_expression.h"
