+++ src/codegen/expression/expression_translator.cpp
@@ -10,7 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "codegen/expression_translator.h"
+#include "codegen/expression/expression_translator.h"
 
 #include "codegen/compilation_context.h"
 #include "expression/abstract_expression.h"
