+++ src/codegen/expression/tuple_value_translator.cpp
@@ -10,7 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "codegen/tuple_value_translator.h"
+#include "codegen/expression/tuple_value_translator.h"
 
 #include "expression/tuple_value_expression.h"
