+++ src/codegen/hash_join_translator.cpp
@@ -14,7 +14,7 @@
 
 #include "codegen/lang/if.h"
 #include "codegen/oa_hash_table_proxy.h"
-#include "codegen/tuple_value_translator.h"
+#include "codegen/expression/tuple_value_translator.h"
 #include "codegen/lang/vectorized_loop.h"
 #include "expression/tuple_value_expression.h"
 #include "planner/hash_join_plan.h"
