+++ src/codegen/translator_factory.cpp
@@ -12,20 +12,20 @@
 
 #include "codegen/translator_factory.h"
 
-#include "codegen/arithmetic_translator.h"
-#include "codegen/case_translator.h"
-#include "codegen/comparison_translator.h"
-#include "codegen/conjunction_translator.h"
-#include "codegen/constant_translator.h"
+#include "codegen/expression/arithmetic_translator.h"
+#include "codegen/expression/case_translator.h"
+#include "codegen/expression/comparison_translator.h"
+#include "codegen/expression/conjunction_translator.h"
+#include "codegen/expression/constant_translator.h"
 #include "codegen/delete_translator.h"
 #include "codegen/global_group_by_translator.h"
 #include "codegen/hash_group_by_translator.h"
 #include "codegen/hash_join_translator.h"
-#include "codegen/negation_translator.h"
+#include "codegen/expression/negation_translator.h"
 #include "codegen/order_by_translator.h"
 #include "codegen/projection_translator.h"
 #include "codegen/table_scan_translator.h"
-#include "codegen/tuple_value_translator.h"
+#include "codegen/expression/tuple_value_translator.h"
 #include "expression/case_expression.h"
 #include "expression/comparison_expression.h"
 #include "expression/conjunction_expression.h"
